[ATLAS] fix(retrieval): break agent_query↔overrides import cycle (CodeQL py/cyclic-import)
Extract Citation to src/retrieval/_types.py (a leaf module importing nothing
from src.retrieval). overrides now imports Citation from _types instead of
agent_query, removing the back-edge that formed the cycle. agent_query
re-exports Citation so existing `from src.retrieval.agent_query import
Citation` callers and the package __init__ are unaffected.

Dependency graph is now a DAG: _types ← agent_query, _types ← overrides,
overrides ← agent_query. The TYPE_CHECKING guard made it runtime-safe, but
CodeQL's structural finding was real — this resolves it at the source.
